+++ Sources/SwiftExtras/Extensions/Date+.swift
@@ -267,6 +267,52 @@ extension Date {
         )
     }
 
+    /// Random date in the current month (can not be in the future)
+    ///
+    /// Returns a random date in the range of dates.
+    ///
+    /// - Parameter range: The range of dates
+    /// - Returns: A random date in the range
+    public static var random: Date {
+        let calendar = Calendar.current
+        let today = Date()
+
+        // Generate a random day within the current day and the first day of the month
+        let randomDay = Int.random(in: 1 ..< Date().day)
+
+        // Create the date with the random day
+        var components = calendar.dateComponents([.year, .month], from: today)
+        components.day = randomDay
+        components.hour = Int.random(in: 0...23)
+        components.minute = Int.random(in: 0...59)
+        components.second = Int.random(in: 0...59)
+
+        return calendar.date(from: components) ?? Date()
+    }
+
+    /// Random date in the current month (can be in the future)
+    ///
+    /// Returns a random date in the range of dates.
+    ///
+    /// - Parameter range: The range of dates
+    /// - Returns: A random date in the range
+    public static var randomInMonth: Date {
+        let calendar = Calendar.current
+        let today = Date()
+
+        // Generate a random day within the current day and the first day of the month
+        let randomDay = Int.random(in: calendar.range(of: .day, in: .month, for: today) ?? 1 ..< Date().day)
+
+        // Create the date with the random day
+        var components = calendar.dateComponents([.year, .month], from: today)
+        components.day = randomDay
+        components.hour = Int.random(in: 0...23)
+        components.minute = Int.random(in: 0...59)
+        components.second = Int.random(in: 0...59)
+
+        return calendar.date(from: components) ?? Date()
+    }
+
     /// End of the day
     public var endOfDay: Date {
         var components = DateComponents()